API conection with data base
